Checkin of the rest of the changes to the user manual by Feliks Kluzniak. As he made numerous changes, and added macros, I am checking this into the main rather than the patch branch.
Email from Feliks about his changes (2009-06-03):
===========
Attached please find the tex files from documents/userman that contain my
(extensive!) suggested improvements.  There is also a new file,
umsmacros.sty, which should reside in the same directory as the tex files
and takes over the role of sepiachiphtml.tex for userman.

This took over a week of concentrated effort, so I guess I won't be doing
this for the other documents anytime soon.  :-(   I split off the new macro
file exactly for that reason.

Some remarks:

- My primary motivation was to improve consistency throughout the document.
Perfection has not been achieved, but it's a start.  I hope it will be
possible to do future modifications in a more piecemeal fashion.  This will
be helped in particular by the fact that I removed almost all the instances
of font changes embedded directly in the text (one should always use
macros!).

- I concentrated mostly on making the printed version look reasonably good.
I feel the HTML turned out not to be too bad, either, except of course for
the crappy small courier font.

- In the interest of consistency I decreed that it will be "built-in", not
"builtin" (you will remember we discussed that), "finalize" and not
"finalise" (though I tend to use the latter in my texts), and "file name",
not "filename" (ditto).

- I have split all occurrences of "egin{quote}egin{verbatim}" into two
lines, because they suppressed the colouring of keywords etc. in the LaTeX
mode of Emacs (a most useful feature!).  For consistency, I did the same for
the "end" sequences.

- I tried to replace all the instances of the Americanism "need to" with
proper English equivalents.  This is a part of my personal crusade against
the monstrosity.

- In some twenty or thirty cases you will see perfectly good formulations
modified for no apparent reason.  This was done to avoid complaints from TeX
about overfull lines.  Please notice that I was able to reduce the number of
warnings from LaTeX from a great many to just two very trivial ones.  (I did
not even attempt to look at the numerous warnings from the other tools: I
ran out of time.)

- I am not sure it is such a great idea to include the explanation of
debugger commands in their index entries: this makes the index look really
bad.  If you agree with me, it's now a matter of massaging a single macro in
umsmacros.sty (dbgcmdidxplus).

================
Chages I have made to what he sent me:

1) Check and included changes to the main branch userman not in the patch branch (which he was using)
2) Reverted his changes of _ to _, as we have a macro to use _.
3) Updated sepiachiphtml.tex to include his macros in umsmacros.sty, which was a modified sepiachiphtml.tex. Checked that this still works with the other manuals.