fixed a problem introduced in the recent rewrite to handle errors: gcc kept failing -- the problem was reported by Thorsten Winterer.
The gcc is failing inside a block/3, and it is not clear why this is so, as the code seem to work without the block/3.

The code is rewritten so that block/3 is not needed for the error handling. The existing code was also more complicated than needed, so it was simplified.